+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+CHANGELOG.md merge=union
+
 ## 2026-08-07
 
 - observability-monitoring-concepts: Add docs combining observability with containerization (con-038)
@@ -7,6 +9,7 @@
 - prom: Add quick primer explaining Prometheus as a time-series metrics monitoring toolkit (prom-001)
 - prom: Add install and web UI exploration notes with first-time impressions (prom-002)
 - prom: Add first Prometheus scrape target configuration (prom-003)
+- observability-monitoring-concepts: Add Python snippet for applying observability in DevOps (con-046)
 
 ## 2026-08-06